PutAssignee: Fail with 400 Bad Request if in input no assignee is set
So far SetAssigneeOp accepted null as assignee if there was no old
assignee, but then trying to add the null assignee as reviewer failed
with a NullPointerException in PostReviewers#prepareApplication(...).

Deleting of assignees should be done via the DeleteAssignee REST
endpoint.

Move the permission check out of SetAssigneeOp, because we always do
this first before checking for bad requests (e.g. see SetHead).

Change-Id: Ia19e7a47d3925215deb13eb78e1814bb42cf5bae
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 9b8e8895774b8393d686a9ecf4a599abdbff50ca