Fix autocomplete popup focus error
Previously, clicking on an autocomplete suggestion was erroneously
determined to be a click outside the autocomplete. This occurred
because the <li> elements that made up the suggestions were completely
detached from the DOM before their hierarchy in the DOM was able to be
checked. This change fixes that by switching the click detection to tap
detection, and stopping the event propagation inside the suggestion tap
handler.

Bug: Issue 4747
Change-Id: I1a3887bca3057ffa9efb581fe0cf2b9d3b156a77

Former-commit-id: 1805cf5d702ccd56e20be90a46c90d7908f20b87