Fix flaky test in gr-autocomplete
The '_focused flag properly set' test in gr-autocomplete failed ~25% of
runs. Eventually, the case was isolated down to whenever the tests were
running as the active tab in one browser window, but that window itself
was not focused. This may have something to do with browser throttling
of function calls in windows that are determined to be inactive.

Whatever the case may be, switching to a flush() call and using
MockInteractions.focus() instead of input.focus() fixes the issue in
every run.

Bug: Issue 4779
Change-Id: I06df1c3573b1024fff22a4033bc3b695c6288331

Former-commit-id: a51a24cbe8004b3afaf2eaa0ef238edadff6467e