Refresh change when approved status changes
With the addition of inline vote deletion, a few different
circumstances arose where the change metadata, including the
submittable state and the labels remaining needed for submission,
could become stale. This change refreshes the change view any time a
vote that may affect the submittable status of the change is removed.

Bug: Issue 4759
Change-Id: I56058126dd13085a3f9ceaf0ead03e30e6d93b83

Former-commit-id: f0743733190022fa4f1ddff430c1fb6df2383bf2