Declare dependency for file list action method
The `gr-file-list#_computeShowFileListActions` method made use of the
`_maxFilesForBulkActions` property, but did not accept it as an
argument. This resulted in bindings executing the compute method
before all of its dependencies were initialized, resulting in the
file list actions being hidden in all cases.

With this change, the dependency is declared and is passed in each use.
The method name is also shortened a bit.

Change-Id: I61a9479a10cdea262b8e717292ba206bf8549bc1

Former-commit-id: d824c682dabc0cc15ed81cf94e11874432e09396