AllProjectsCreator: Allow admins and owners to create refs/meta/config
Nowadays when a project is created in Gerrit it always has a
refs/meta/config branch, but the refs/meta/config branch may be
missing for imported projects. In this case it is good if the Gerrit
administrators and the project owners can just create the missing
refs/meta/config branch.

Add a schema migration that grants the 'Create Reference' access right
on refs/meta/config to all administrator groups and project owners in
All-Projects.

Also old projects may miss the refs/meta/config branch and having this
'Create Reference' permission on refs/meta/config assigned makes it
easy to add the missing refs/meta/config branch for these projects.

Change-Id: I634bd21189196b479eb6d4b578d684bc9b54d49c
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 541d9747d73922b105b7a59e53e99bd003ffa3a6