Refactor long method into smaller ones
In order to improve readability, several smaller methods are
introduced which encapsulate similar aspects. To achieve the latter
one, the reviewed state of patchsets is updated at a later point of
time than before. This has the additional benefit that the state will
only be updated if the change and all patchsets are deletable.

Change-Id: I9b631a8dd74b7c38f7a37d622d615220a05ee738

Former-commit-id: 7ddab122b35c9f306c7e8c34fda3a1d5991a7e43