GroupList: Do not store null UUIDs
A GroupList is a parsed representation of the groups file, which
cannot contain any null entries. Caching null entries in this class is
not appropriate.

In addition to this theoretical objection, it is definitely incorrect
to return a cached GroupReference with null UUID from #resolve.
ProjectConfig will create a distinct GroupReference with a different
name for each group that is missing in the groups file, and it's not
correct to return a GroupReference from resolve with a different name
from the one that was passed in, even if they both have null UUID.

This commit additionally reverts 26fe49b, since the new implementation
of resolve would just return the input anyway, so the call would be a
no-op.

This reverts commit 9baef10113c1ac5f5f025d994dd12228b4eb2b17 [formerly 26fe49b72f6075fdc36dbdcfbbc5c878cbcd8c14].

Change-Id: I27e4ed5443122557ab3a719b8bdb052a5e373f58

Former-commit-id: 32f547596cf8553c2b7fd209195c34886010aa15