ProjectCacheImpl: Harden against null UUIDs
Since Ie51b350d GroupList#uuids may include null UUIDs. Regardless of
whether or not this is a good idea, these UUIDs are never relevant as
defined by ProjectCache#guessRelevantGroupUUIDs. Harden
ProjectCacheImpl against this state of affairs as well as future bugs.

This is a backport of I74c62a500 from master, which cannot be directly
cherry-picked to stable-2.13 due to using Java 8 specific syntax. This
is re-implemented in a Java 7 compatible way.

Change-Id: I448e3f30374ebf25c44f3b3a6911db5725173a28

Former-commit-id: 84acf97c788117d1a74c1a18ad733198e8129418