Add new submit strategy "Rebase Always".
This strategy works the same way as Rebase If Necessary except that it
inserts new patchset to every change even if fast foward if possible,
in which case new patchset's only difference is commit message. Hence,
Rebase Always provides all advantages of Rebase If Necessary.

Adding extra Reviewed-By and Reviewed-On footers as Cherry Pick will
be implemented in the follow up CLs.

Bug: Issue 4452
Change-Id: I8c2514c69c404eebca3469e3f49a51793d98ab95

Former-commit-id: a11e63df5cdc1c2a362b355d471fac210313eccc