Bazel: Simplify diffy_logo rules
There is a regression in Bazel 0.3.2 release: [1] that prevents creating
a symbolic link to previously built artifact. In this case we don't
really need to do that, as we can depend on previously built library and
do not need to create a zip from java_library and import it again as
java_library.

[1] https://github.com/bazelbuild/bazel/issues/1966

Change-Id: I16a62e2fbb3e2d1de377eeef70061298bb2bb745

Former-commit-id: 0e121132f51c25acba760509404f8dfe00b98314