Prevent circular module dependency when running in external container
Commit 220a3f5 introduced circular dependency between the database
module and NotesMigration (impl. class ConfigNotesMigration).

Having the site path being stored in the database was broken by above
change. The class Daemon solves this by explicitly setting the site
path and adds a SitePath module before starting everything.

This patch makes it possible again to run the gerrit war in an external
Servlet container, but you need to set the System property
"-Dgerrit.site_path=/path/to/gerrit" to make it work.

The site path stored in the database became practically obsolete with the
NotesMigration change mentioned above.

Change-Id: Id735fbf928dda186e659792c84d4e675c893a29a

Former-commit-id: d66f37c5ab73049383a5dabbd1aa4691433eaa7c