Fix for autolinking missing matches
Previously, there was an issue where text would not get autolinked if
there were multiple regex patterns and a match for one was found below
the occurance of another that was not checked for yet.  This change
addresses the issue by checking the full text block for matches for each
type of regex, regardless if one has been found above it.

Bug: Issue 4325
Change-Id: I11db954de6a1da26b8ab29b8501b1f1f01425364

Former-commit-id: 423f298512584cdf1aeed2324c60ffb89fd54c92