ChangeNotestTest: Always add new patch set
Some tests were incrementing the currentPatchSetId field in the
Change, and possibly referring to it from some dependent entities in
the test, but not actually populating the patch set with a Commit
footer. Future changes to the parser will affect the behavior in these
cases, so the tests need to create some more realistic data.

Change-Id: I9a91fae24d22a23df7659d408a8a3d895bcb8594

Former-commit-id: 6e0d5dc61693aaf9dd76c001950b7749ed0ed698