ReplaceOp: Delete dead code
The change.getStatus().isClosed() check in resetChange was already
done at the very top of updateChange, so the second case is dead code.

It's a good thing it was dead code, because it didn't come close to
undoing all the mutations that had happened so far, and didn't cause
the transaction to abort.

Change-Id: If1fccba0d1b8771734fb60fb3cdd6403f624f855

Former-commit-id: ce60204753d6d93df01b1e1aaa2a0a101a139981