AbstractQueryChangesTest: Create changes at explicit timestamps
The old implementation of these test methods used whatever
TimeUtil.nowTs() would report as the current time, so the queries were
dependent on how many times TimeUtil.nowTs() was invoked in the change
creation path. A future change will make that specific behavior
different between ReviewDb and NoteDb, which would break these tests.

Rewrite the tests to explicitly set the timestamp for each change to
the value that the test expects.

Change-Id: I326e804b566eb1b171801f465ba8b88d1467b184

Former-commit-id: 782cf4b0e4e0a5f448b6a0d4926f7a69fdd88691