Add check before requesting conflicts in related change view
Previously, an API request for getting conflicts was made no matter
what. This request was often very slow, and because the related change
element waits for many promises to resolve before rendering, it was a
bottleneck to display changes. In the GWT interface, that particular
call is only made if a change is open and mergable. With this change,
the same kind of checks have been added to Polygerrit in order to make
the relation chain appear faster in many cases.

Bug: Issue 4727
Change-Id: I9e0b42265a863c86de5eec3328b37e222eba2f97

Former-commit-id: 04d5a3f2ef5a4b6a470b80a730d739cd884a4e34