SetMembersCommand: Handle REST errors gracefully
If a REST error occurs while adding a member or group to a group, the
command fails with "internal server error" which is not helpful to the
user.

Catch REST errors and allow them to be reported back to the user.

Other types of error will still result in "internal server error".

Change-Id: Ia21f7aa28212f6d93ec4e6ff74a9de9ee6428a00

Former-commit-id: 664a284bc6bdfed6c07b31ad0ae597bf7c1ff36c