ReviewersUtil: Query for groups only while limit is not reached yet
It is unnecessary to query for groups to suggest if they will be
dropped later because the limit was already reached.

Extend SuggestReviewersIT so that we test the correct behavior when
the suggest limit is exhausted before we start to suggest groups and
when the suggest limit is exhausted while we are in the middle of
suggesting groups.

Change-Id: Ifa04ccb136abbfe2c155e148b33776a78056e688
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 9b6b424a391cbe9a4059e4aa74ecd526bbaaf60f