Strip servlet API from GWT jar to avoid classpath collision
gwt-user.jar is packaging javax.servlet stuff in older version that our
own servlet_api_3_1. In case plugins depend on the new API they get
compilation failures that hard to track down und understand. We are
already stripping outdated Jetty stuff from gwt-dev.jar artifact so that
we don't mind to do it also for gwt-user.jar. The good news are in GWT
2.8 the most dependencies were externalized.

Reported-by: David Pursehouse <dpursehouse@collab.net>
Change-Id: Ief44bc676fca9bcd2fb1e4f7dff4547d1604c5cd

Former-commit-id: bb4fca2732df2193cf92c7e8f2e5e581ea695d1f