PolyGerrit: Don't create development servlets in production code path
To satisfy Guice wiring, it's not necessary needed to create a dummy
provider that is not actually used. Use @Nullable instead to bind
to Provider that optionally returns null. This simplifies the code
in development servlets and avoid eager created (unused) development
instances in production code path.

Also rename the development servlets with Dev suffix to make the intent
more clear.

This is a preparation change to abstract the build system and support
Bazel in addition to Buck.

Change-Id: I9c2a7b61509de446e292af9cb0217d24547bbbf5

Former-commit-id: 5a09d99adddf20b7c0920a12f732b69fc5dc6182