Fix use of closed Repository in CommentSender
CommentSender::getGroupedInlineComments returns PatchFile objects that
hold a reference to a temporarily opened Repository. These objects may
later be passed to CommentSender::getLine, which in turn attempts to
make use of the Repository to read the line from the file that
corresponds to a particular comment.

This change moves management of the Repository's lifetime up the stack,
so that it remains open for the duration of compiling all the comment
details.

Change-Id: I898d0f4e0485fbfe905654eccb65aca57fdeb1ca

Former-commit-id: a46f9e510dff87a28fa5af50ee50cade4484c5fe