Fix issue where keyboard shortcuts don't appear
Previously on a change view, if you click either the expand all or the
hide automated message links (and don't click elsewhere on the page),
keyboard shortcuts don't work.  This was due to the focus still being
on the link, which doesn't trigger the keyboard shortcuts.

This change removes focus on the links after they are clicked so even if
the user doesn't click elsewhere on the page, keyboard shortcuts will
still work.

Bug: Issue 4888
Change-Id: I51527325ea8d5d575c4d3f4ad02bf0b121fa8cce

Former-commit-id: dc41fe2610d310063d287032cd8a2b0c1ca3da5f