Allow submit of merge of non change branch
Commit b3bbcf5a4512cc4e92cb3b93dc301ce3470bb362 [formerly a61b6eed5337049a7feb58ac935543624f107129] introduced a bug that
prevents mergeing of a branch that is not a change.
It marks the branch to be merged into as uninteresting whereas the
branch that gets merged is not, which causes Gerrit to think that it
is not allowed to merge it resulting in MISSING_DEPENDENCIES error.

In the case where the walk hits a commit that does not have canMergeFlag
and is not in the commits to be submitted, check if commit is already
merged.

Bug: Issue 4930
Change-Id: Ib4f95ebe336e381d6fcbf36867e84d927ac13eb1

Former-commit-id: 29445062ad2755a0806985073050423e5e75850f