Remove an obsolete code comment from the AsciiDoctor class
In case of the bazel build no Zip file for the HTML documentation is
created. Instead, we want keep the individual HTML files. With that in
mind it is no longer true that we would like to render HTML to a string
that gets directly added as a ZipEntry, so remove that comment.

Change-Id: I8384c8500e3658aebd6143fbe766cf1944bb4809

Former-commit-id: a111bbc62a08b90602b9505672060248121ddeb4