Bazel: Fix eclipse classpath generation on first build
For GWT SDM mode to work, we need not only libEdit.jar but also
libEdit-src.jar. However in Bazel it is only get built, when it
is explicitly invoked. That why we have to explicitly build this
target. Also remove the assert, because the eclipse classpath
generation happens before the actual build (to prevent having
empty .classpath when compilation errors exist).

Change-Id: I7ec51f1d9e3488742964c49a3d34b46cad4e6bc2

Former-commit-id: 78fa1c8ba455f1290db878c125414247451f44a4