Bazel: Respect test dependencies for classpath generation
Test dependencies must be respected during classpath generation because
some third party dependencies can be only used for the tests, but not
for production code path. Otherwise, we would end up producng classpath
that missing some dependencies and thus compilation errors in the IDE.

This was the case with jimfs, that was added as implicit dependency to
the 'tools/eclipse:classpath' rule.

Skip Test_runner_deploy.jar library from the Eclipse classpath, as it
includes some other third party dependency (most notably outdated
auto-value) that could collide with our own version of those
dependencies, causing classpath collisions, see: [1] for the glory
details: [1]. This is safe thing to do, as we rely on the Eclipse as
JUnit test execution environment anyway.

* [1] https://github.com/bazelbuild/bazel/issues/2044

Change-Id: I87fff277695a2f64c44a3af65471c0c901860a02

Former-commit-id: 7f95f255c1bf05841f81ba980df4aaf401342f46