RebaseAlways does not handle identical tree properly
In RebaseAlways if a commit can be fast forwarded, it will still modify
the commit message, however, MergeUtil.createCherryPickFromCommit can
throw a MergeIdenticalTreeException, in this case, we will not get a
new commit, and will fail at sanity check in
SubmitStrategyListener.findUnmergedChanges.

Note that RebaseIfNecessary will behave different by moving the branch
tip to the commit, it also make sense to make an empty rebase commit and
move the branch to it in RebaseAlways case.

Change-Id: Ia00b32a565031130534a3e25cf27d16af18c850f

Former-commit-id: 9d16b6b0e3e62a40c8c73c8ebcbecb6f9779bb44